refactor!: Un-fork base class and make a silex prototype class
